Move unification into Surface.Unification.
